### 2023-07-25 4e3fbbe
Update paper with suggested DOI in bib and better diagram
