### 2023-07-25 d2eb5ed
Add JOSS badge to readme
