### 2023-07-25 cdd88c5
Update first walkthrough notebook
