### 2023-07-25 823302f
Update example notebook 1 with new terminology
