### 2023-07-26 6574ca7
Finish updating getting_started
