+++ sphinx/source/components.rst
@@ -32,20 +32,20 @@ are simply an alternative way to running stages than the functional form.
         test_model
     ], ...)
 
-**Parameter sets** - Parameter sets represent a specific collection (list of argsets) of configuration values or adjustable parameters for the stage functions
+**Parameter sets** - Parameter sets represent a collection of configuration values or adjustable parameters for the stage functions
 to apply to the functions called in the codebase.
 
-**Records** - A record is a collection of data/variables used in a single run of a procedure with a single paramset.
+**Records** - A record is a collection of data/variables used in a single run of a procedure with a single parameter set.
 A record is automatically passed between stages of a procedure in order to persist the information. The currently
 active record is passed into every stage, so a stage has manual access to everything stored in it.
 
 **Artifact Manager** - An artifact manager instance is associated with each experiment run, and it tracks and handles all intermediate
 outputs from the various stages. The manager is the overarching collection of records for an experiment run.
-It tracks all records for each procedure/paramset
+It tracks all records for each procedure/parameter set
 used. A stage with an aggregate decorator can get access to all records instead of just the currently active one in
 the procedure in order to aggregate results or other data across the records.
 
-**Experiments** - Experiments apply given sets of parameters to one or more procedures, potentially running additional aggregate stages
+**Experiments** - Experiments apply given parameter sets to one or more procedures, potentially running additional aggregate stages
 to collect and analyze results from multiple procedure runs across parameter sets. The intention of an experiment is
 to allow easy comparison of results for multiple similar procedures and/or multiple parameter configurations across one
 or more procedures.
@@ -61,13 +61,13 @@ Stage flow
 .. figure:: images/interactions_stage.png
     :align: center
 
-A stage is the processing of a record. A record has a :code:`state` attribute, which is a dictionary
+A stage is the processing or transformation of a record. A record has a :code:`state` attribute, which is a dictionary
 of data from previous stages. A stage which requires input automatically finds those input variables from the passed :code:`record.state` and
 passes them into the function,
-and the outputs are stored from the returns back into :code:`record.state`. A record is initialized with some :code:`Args` parameter
+and the outputs are stored from the function returns back into :code:`record.state`. A record is initialized with some :code:`ExperimentParameters` parameter
 set, and the record itself has an associated artifact manager.
 
-Note that the four components here, args, records, stages, and a manager are the minimum components required
+Note that the four components here, parameters, records, stages, and a manager are the minimum components required
 to interact with Curifactory. Procedures and experiments are helpful but not required, and you could for
 instance run stages on their own with just these four components in a jupyter notebook or similar.
 (However, most of the benefit of Curifactory comes from the experiment management side through the
@@ -80,13 +80,34 @@ Procedure flow
     :align: center
 
 Procedures simplify calling multiple stages. A procedure is defined with a list of these stages, and then
-calling :code:`proc.run(args)` while passing in an :code:`Args` parameter set will automatically create an associated
+calling :code:`proc.run(params)` while passing in an :code:`ExperimentParameters` parameter set will automatically create an associated
 record, store it in the manager, and then execute the sequence of stages as a pipeline on the record.
 The output is the final record, which could also be accessed as :code:`manager.records[-1]`, since it was the
 last record added to the manager.
 
 Note that the :code:`run()` function optionally takes a record as a parameter, which if passed will use the existing record
 rather than creating a new one, meaning multiple procedures can technically be chained.
+
+
+Any aggregate stages that are applied to a given record must be the first stage in the chain for that record.
[…]
