+++ sphinx/source/parameters.rst
@@ -5,68 +5,70 @@ Another goal of Curifactory is to allow effective parameterization of experiment
 done with a json or yaml file, Curifactory directly uses python files for experiment parameterization/configuration.
 This has a few advantages:
 
-1. Arguments can be any python object, rather than simply a primitive type or dictionary.
+1. Parameters can be any python object, rather than simply a primitive type or dictionary.
 2. Parameter files can reference/use other parameter files, allowing modularity and composition.
-3. The resulting arguments that are passed into an experiment can be algorithmically generated or modified inside an
-   arguments script file, with the full power of the python language! An example for how this might be useful is a single
-   arguments script that generates 10 very similar argument sets for comparison, rather than having to individually define
+3. The resulting parameter sets that are passed into an experiment can be algorithmically generated or modified inside an
+   parameter file, with the full power of the python language! An example for how this might be useful is a single
+   parameter file that generates 10 very similar parameter sets for comparison, rather than having to individually define
    10 different parameter configuration files. This could allow custom gridsearches for example.
 
 .. note::
 
-    Throughout this documentation, we refer to "paramset" and "argset" as slightly different. A **"paramset"** refers
-    to a whole parameters script file, while an **"argset"** refers to a single :code:`Args` instance. A single paramset
-    returns one or more argsets in a list.
+    Throughout this documentation we use specific language to refer to different parts of parameterization:
 
-The :code:`Args` class
-----------------------
+    * **Parameter** - a single parameter is a single attribute of a parameter class.
+    * **Parameter class** - a dataclass that extends curifactory's ``ExperimentParameters`` and defines the possible hyperparameters.
+    * **Parameter set** - an instance of a parameter class, every stage operates on a record with a single specific parameter set.
+    * **Parameter file** - a python script as defined below that creates one or more parameter sets.
 
-As discussed on the :ref:`Getting Started` page, To define possible arguments, there should be a
-class that inherits :code:`curifactory.ExperimentArgs`, and for ease of use should have the
-:code:`@dataclass` decorator. Possible arguments are the variables within
-this class, and by defining default values for each one, this allows an arguments file to define
-only what it needs to change from the defaults.
+The :code:`ExperimentParameters` class
+--------------------------------------
 
-An example :code:`Args` class is shown below:
+As discussed on the :ref:`Getting Started` page, To define possible parameters, there should be a
+class that inherits :code:`curifactory.ExperimentParameters`, and for ease of use should have the
+:code:`@dataclass` decorator. Possible parameters your experiment stages can use are the attributes
+within this class, and by defining default values for each one, a parameter set constructor need only
+specify the parameters that differ from the defaults.
+
+An example :code:`Params` class is shown below:
 
 .. code-block:: python
 
     from dataclasses import dataclass, field
-    from typing import List
 
-    from curifactory import ExperimentArgs
+    from curifactory import ExperimentParameters
 
 
     @dataclass
-    class Args(ExperimentArgs):
-        example_arg: str = ""
+    class Params(ExperimentParameters):
+        example_param: str = ""
         example_number_of_epochs: int = 10
 
         # due to how dataclasses handle initialization, default lists and dictionaries need to
         # be handled with field factory from the dataclasses package.
-        example_data: List[int] = field(default_factory=lambda: [1,2,3,4])
+        example_data: list[int] = field(default_factory=lambda: [1,2,3,4])
 
 
[…]
