### 2023-07-27 370f8a4
Expand caching documentation
