+++ sphinx/source/cache.rst
@@ -1,33 +1,58 @@
 Cache
 =====
 
-Including a list of cachers in your stage decorators helps store intermediate results
-both for easier exploration as well as faster re-execution.
+Curifactory makes it straightforward to store and re-use intermediate artifacts generated
+throughout an experiment with its caching mechanisms. During an experiment run, user-specified
+caching strategies dump parameter-set-versioned instances of stage outputs in a common cache folder,
+and when running a stage that already has the appropriate artifacts in the cache for the current
+parameter set, it uses the caching strategy to reload the artifact from cache instead of executing
+the stages. Storing artifacts in cache both helps re-execute the experiment faster as well as
+creates a "paper trail" for manual exploration of the artifacts.
 
-There are four pre-implemented cachers that come with Curifactory in the :ref:`Caching`
-module that should cover most basic needs:
+Caching strategies are ``Cacher`` classes that extend curifactory's base ``Cacheable`` class. Using
+these cachers is as easy as listing them in your stage decorator for each output the stage generates:
 
-* :code:`JsonCacher`
-* :code:`PandasCsvCacher`
-* :code:`PandasJsonCacher` - stores a dataframe as a json file (array of dictionaries, the keys as column names.)
-* :code:`PickleCacher`
+.. code-block:: python
+
+    from curifactory import stage
+    from curifactory.caching import PandasJsonCacher, JsonCacher, PickleCacher
+
+    @stage(outputs=["dataset", "metrics_dictionary", "model"], cachers=[PandasJsonCacher, JsonCacher, PickleCacher])
+    def return_all_the_things(record):
+        ...
+        return dataset, metrics, model
+
+
+There are several pre-implemented cachers that come with Curifactory in the :ref:`Caching`
+module that should cover many basic needs:
+
+* ``JsonCacher``
+* ``PandasCsvCacher``
+* ``PandasJsonCacher`` - stores a dataframe as a json file (array of dictionaries, the keys as column names.)
+* ``PickleCacher``
+* ``FileReferenceCacher`` - a json file that stores references to one or more file paths.
+* ``RawJupyterNotebookCacher`` - turns a list of list of strings of python code into a jupyter notebook
 
 As a last resort, most things should be cacheable through
-the :code:`PickleCacher`, but the advantage of using the :code:`JsonCacher` where
+the ``PickleCacher``, but the advantage of using the ``JsonCacher`` where
 applicable allows you to manually browse through
 the cache easier, instead of needing to write a script to load a piece
 of cached data before viewing it.
 
-Some things may not cache correctly even with a :code:`PickleCacher`,
+Some things may not cache correctly even with a ``PickleCacher``,
 such as pytorch models or similarly complex objects. For these, you
 can write your own "cacheable" and plug it into a decorator in the same
 way as the pre-made cachers.
 
 Implementing a custom cacheable requires extending the :class:`caching.Cacheable <curifactory.caching.Cacheable>`
-class, and the new class must have a :code:`load()` and :code:`save()`
-function. The base class has a :code:`path` attribute that both functions can assume
-is set correctly to a base path where it is appropriate to write any necessary files.
-Following is an example:
+class, and the new class must have a ``save(obj)`` and ``load() -> obj``
+function, which respectively should handle saving the passed artifact to disk,
+and loading and returning a reconstructed artifact.
+
+The base ``Cacheable`` has a ``get_path()`` function which the cacher implementation can assume
+correctly returns a full filepath including the correct versioned filename for the current artifact.
+In the case that a cacher needs to save more than one file or wants to provide a different suffix for
+the filename, this can be passed to ``get_path``.
 
 .. code-block:: python
 
@@ -35,27 +60,43 @@ Following is an example:
     from curifactory.caching import Cacheable
 
     class TorchModelCacher(Cacheable):
-        def __init__(self):
[…]
