### 2023-07-27 9241213
Fix missing extension None check on cacher
