+++ sphinx/source/cookbook.rst
@@ -25,3 +25,9 @@ Using slurm in HPC environments
 ===============================
 
 TODO
+
+
+Cachers for Pytorch Lightning models
+====================================
+
+TODO
