### 2023-07-27 ee47ea4
Update report args section headings
