+++ sphinx/source/config.rst
@@ -43,12 +43,15 @@ The default values for the configuration are shown in this example:
         "report_css_path": "reports/style.css",
     }
 
-:code:`experiments_module_name` - the name of the folder where experiment
+:code:`experiments_module_name` - The name of the folder/module where experiment
 scripts are stored. This is treated as a python module, running an experiment
-essentially runs :code:`import experiments.[experiment_script_name]`
+essentially runs :code:`import experiments.[experiment_script_name]`. This means
+you can submodule your experiments folder. Note that since this is a module name,
+if you have it in a subfolder be sure to use '.' in your config instead of '/'
 
-:code:`params_module_name` - the name of the folder where parameter scripts
-are kept.
+:code:`params_module_name` - The name of the folder/module where parameter scripts
+are kept. Similar to experiments, this is a module name, and you can submodule
+your parameters.
 
 :code:`manager_cache_path` - The folder where artifact manager data is kept,
 namely the experiment store.
