### 2023-07-27 2849267
Update example experiments to use new params
