### 2023-07-27 42ae33b
Reorganize examples folder
Closes #61
