### 2023-07-28 dae23a6
Add hashing mechanics documentation
