### 2023-07-30 8e87011
Update tips and cli docs
