+++ sphinx/source/cli_guide.rst
@@ -1,7 +1,7 @@
 CLI Guide
-=============
+#########
 
-The primary means of running Curifactory is through the provided :code:`experiment` CLI.
+The primary means of running Curifactory is through the provided ``experiment`` CLI.
 The three main tasks that can be run through this command line tool are:
 
 * Run an experiment.
@@ -9,22 +9,22 @@ The three main tasks that can be run through this command line tool are:
 * Serve the local report HTML files via python simple web server.
 
 Running an experiment
----------------------
+=====================
 
-Running the :code:`experiment` tool always checks for a local :code:`curifactory_config.json` for
+Running the ``experiment`` tool always checks for a local ``curifactory_config.json`` for
 relevant paths, see :ref:`Configuration`.
 
 Running an experiment requires the name of an experiment file (minus the .py)
-and at least one parameters file (minus the .py) via the :code:`-p` argument. Note that the filenames
+and at least one parameters file (minus the .py) via the ``-p`` argument. Note that the filenames
 you pass are not full paths, they are only filenames within the paths specified in the
-configuration (by default an :code:`/experiments/` and :code:`/params/` folder.)
+configuration (by default an ``/experiments/`` and ``/params/`` folder.)
 
 .. code-block:: bash
 
     experiment [EXPERIMENT_FILE] -p [PARAMS_FILE]
 
-The above command calls the :code:`get_params()` from the :code:`PARAMS_FILE` and passes the returned list
-of :code:`Args` instances into the :code:`run()` function inside :code:`EXPERIMENT_FILE`.
+The above command calls the ``get_params()`` from the ``PARAMS_FILE`` and passes the returned list
+of ``ExperimentParameters`` instances into the ``run()`` function inside ``EXPERIMENT_FILE``.
 
 For example, with the following project setup:
 
@@ -43,7 +43,7 @@ We would run the experiment from the project root directory with:
     experiment compare_models -p sklearn_algs
 
 
-You can optionally specify multiple parameter files, where the combined returned :code:`Args`
+You can optionally specify multiple parameter files, where the combined returned ``ExperimentParameters``
 lists from each file are placed into a single list and passed into the experiment.
 
 .. code-block:: bash
@@ -52,10 +52,10 @@ lists from each file are placed into a single list and passed into the experimen
 
 
 Curifactory checks both the params module path as well as the experiments module
-path for files with valid :code:`get_params()` files, so an experiment file that
-has this function may also be used in a :code:`-p` flag.
+path for files with valid ``get_params()`` files, so an experiment file that
+has this function may also be used in a ``-p`` flag.
 
-If you wish to run an experiment that has a :code:`get_params()` defined and you
+If you wish to run an experiment that has a ``get_params()`` defined and you
 want to use it as your parameter set, you can simply use
 
 .. code-block:: bash
@@ -69,106 +69,106 @@ Which curifactory will automatically expand to
     experiment [EXPERIMENT_FILE] -p [EXPERIMENT_FILE]
 
 
-Filtering to specific args
---------------------------
+Filtering to specific parameter sets
+====================================
 
-When there are parameter files that return large lists of :code:`Args` instances, it can be useful
+When there are parameter files that return large lists of ``Params`` instances, it can be useful
 to run only a small subset of them to update specific results or for testing purposes.
 
-This can be done with any of the :code:`--names`, :code:`--indices`, or :code:`--global-indices` arguments.
+This can be done with any of the ``--names|-n``, ``--indices``, or ``--global-indices`` arguments.
 
 For the following examples, assume we have a single parameters file returning three argsts:
 
 .. code-block:: python
     :caption: Example parameters file
 
-    def get_params() -> List[Args]:
+    def get_params() -> List[Params]:
         return [
-            Args(name="baseline_knn", ...)
-            Args(name="baseline_svm", ...)
[…]
