+++ sphinx/source/tips.rst
@@ -8,24 +8,24 @@ Directly/manually running stages can be useful for a variety of reasons, whether
 using the same configuration values as from an experiment, or as a basis for exploration in
 a jupyter notebook.
 
-The smallest set of components needed to do this is an :code:`Args` instance, an
-:code:`ArtifactManager`, and a :code:`Record`:
+The smallest set of components needed to do this is an ``Params`` instance, an
+``ArtifactManager``, and a ``Record``:
 
 .. code-block:: python
 
     # required components
     from curifactory import ArtifactManager, Record
-    from params import Args
+    from params import Params  # a custom dataclass extending cf.ExperimentParameters
 
     # import whatever stages you need to run
     from stages.data import get_data, clean_data
 
     # initialize instances of the required components
-    args = Args() # create/initialize the args. Note that you could also directly load
-                  # and manually call the get_params() of a paramset script
+    params = Params() # create/initialize the params. Note that you could also directly
+                      # load and manually call the get_params() of a parameter file
     mngr = ArtifactManager("some_cache_ref_name") # the caching prefix name (normally
                                                   # the experiment name)
-    record = Record(mngr, args)
+    record = Record(mngr, params)
 
     # run desired stages
     clean_data(get_data(record))
@@ -33,75 +33,75 @@ The smallest set of components needed to do this is an :code:`Args` instance, an
     # get the return from the last run stage (you could also get it from record.state)
     print(record.output)
 
-Note that calling any stage, e.g. :code:`get_data(record)` returns the modified record that
+Note that calling any stage, e.g. ``get_data(record)`` returns the modified record that
 was passed in, and since calling a stage always only takes exactly one parameter, (the record
-to use) you can directly chain the stages together with :code:`clean_data(get_data(record))`
+to use) you can directly chain the stages together with ``clean_data(get_data(record))``
 
 
 Common dataset for multiple different argsets
 ---------------------------------------------
 
-For some experiments, you may have a single dataset processing procedure that all other argsets should
-share. By default, caching only applies to a single argset, so the data would have to be recomputed for
-every argset, if you defined your entire experiment in one procedure. However, it's possible to write
+For some experiments, you may have a single dataset processing procedure that all other parameter sets
+share. By default, caching only applies to a single parameter set, so the data would have to be recomputed for
+every parameter set, if you defined your entire experiment in one procedure. However, it's possible to write
 your experiment to "share" the state of a data processing record with the remainder of your argsets.
 
 .. code-block:: python
 
-    def run(argsets: List[Args], mngr: ArtifactManager):
+    def run(param_sets: List[Params], mngr: ArtifactManager):
 
         # define the procedures
         common_data_proc = Procedure([get_data, clean_data], mngr) # the resulting state of this procedure
                                                                    # should be the same for all our
-                                                                   # argsets, so we only want to run it
-                                                                   # once
+                                                                   # parameter sets, so we only want to run
+                                                                   # once it
         model_proc = Procedure([train_model, test_model], mngr) # this procedure is where we expect our
-                                                                # argsets to differ, so run it per argset
+                                                                # argsets to differ, so run it per param set
 
-        # run our common procedure once (assuming all of the relevant args are _actually_ the same, we can
-        # just pull from the first one)
-        data_record = common_data_proc.run(argsets[0])
+        # run our common procedure once (assuming all of the relevant parameter sets are _actually_ the
+        # same, we can just pull from the first one)
+        data_record = common_data_proc.run(param_sets[0])
 
         # run each argset through model procedure
[…]
