+++ sphinx/source/cookbook.rst
@@ -1,33 +1,196 @@
 Curifactory Cookbook
 ####################
 
-TODO
+This page contains a set of recipes or solutions to challenges we've encountered.
+
 
 Distributed ``torchrun`` via external stage
 ===========================================
 
-TODO (using click/torchrun etc.)
+Training pytorch models in distributed mode can sometimes cause issues in curifactory - when running
+on GPUs with DDP, pytorch will start a new process for each GPU and run a copy of whatever command
+initially spawned it. If directly setting up a pytorch lightning trainer inside a stage, this means
+that each process will be re-running the entire curifactory experiment.
+
+In theory this should be fine - curifactory can detect the MPI env variables and put itself into
+``--parallel-mode``, and as long as stages are set up appropriately any work prior to the trainer
+should already be cached anyway. However, in practice we've run into performance issues.
+
+A pattern that works relatively well in curifactory to get around this is having a stage act simply
+as a wrapper around an external command - using either ``os.system`` or ``subprocess.run`` (or even
+curifactory's own internal utility for this, ``curifactory.utils.run_command``) to run a separate
+python script (or any other CLI tool, this can be a way to integrate with other languages and
+applications), and then passing in parameters as CLI args, artifact inputs as paths, and desired
+output locations as paths.
+
+The stage then handles getting the associated paths of everything, constructing the CLI line, running it,
+and then loading any outputs back into curifactory and returning them.
+
+In particular, we've found it pretty easy to use the click python library to quickly make an external
+python function into something callable by a stage.
 
 
-Model checkpoint resuming
-=========================
+.. code-block:: python
+    :caption: external/my_model_trainer.py
 
-TODO
-
-
-Interactive jupyter notebook artifact
-=====================================
-
-TODO
+    import click
+    from pytorch_lightning import Trainer
+    from pytorch_lightning.callbacks import ModelCheckpoint
 
 
-Using slurm in HPC environments
-===============================
+    @click.command()
+    @click.option("--dataset_path", type=str)   # input artifact path
+    @click.option("--model_path", type=str)     # output artifact path
+    @click.option("--checkpoint_dir", type=str)
+    @click.option("--num_gpus", type=int)       # a parameter
+    def train_model(dataset_path, model_path, checkpoint_dir, num_gpus):
+        # load data from dataset_path however needed
+        ...
 
-TODO
+        checkpoint_callback = ModelCheckpoint(
+            dirpath=checkpoint_dir, verbose=True, save_last=True, every_n_epochs=1
+        )
+
+        trainer = Trainer(
+            default_root_dir=checkpoint_dir,
+            enable_checkpointing=True,
+            callbacks=[checkpoint_callback],
+            accelerator="gpu",
+            devices=num_gpus,
+            strategy="ddp",
+        )
+
+        # create model
+        ...
+
+        trainer.fit(model, data, ...)  # pass in data however needed, e.g. if
+                                       # using a datamodule
+
+        model.trainer.save_checkpoint(model_path) # put a final copy of the model
+                                                  # at expected output location
+
+        if __name__ == "__main__":
+            train_model()
+
+
+On the curifactory side, it's currently a little annoying to get the correct paths
+for input artifacts (this will hopefully change in a couple versions). Essentially
+you either need to store the string path directly as an artifact, or you can use
+a lazy cacher with ``resolve=True`` (meaning when used as an input, you get access
+directly to the ``Lazy`` object, thus the cacher, thus the cacher's ``get_path()``)
+
+An example of how you would set this up in some prior stage:
+
+.. code-block:: python
+    :caption: stages/data_setup.py
+
[…]
