### 2023-07-30 e6e01cd
Create PandasCacher abstraction for more IO types
