### 2023-07-31 aeab42e
Fix deprecation warnings in tests
