+++ test/test_dag.py
@@ -7,8 +7,8 @@ from curifactory.caching import PickleCacher
 def test_child_records_empty_for_blank_records(configured_test_manager):
     """Records with no relation to eachother should have empty child_records"""
     configured_test_manager.map_mode = True
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
-    r1 = cf.Record(configured_test_manager, cf.ExperimentArgs("test1"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
+    r1 = cf.Record(configured_test_manager, cf.ExperimentParameters("test1"))
 
     configured_test_manager.map_mode = False
     configured_test_manager.map_records()
@@ -29,9 +29,9 @@ def test_child_records_after_make_copy(configured_test_manager):
         t = 5 + 2
         print(t)
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
     r0 = do_thing(r0)
-    r1 = r0.make_copy(cf.ExperimentArgs("test1"))
+    r1 = r0.make_copy(cf.ExperimentParameters("test1"))
 
     configured_test_manager.map_mode = False
     configured_test_manager.map_records()
@@ -54,7 +54,7 @@ def test_output_used_check_when_no_following_stages(configured_test_manager):
     def thing1(record):
         return "green eggs and ham"
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
     r0 = thing1(r0)
 
     configured_test_manager.map_mode = False
@@ -74,10 +74,10 @@ def test_output_used_check_when_record_copied_butunused(configured_test_manager)
     def thing1(record):
         return "green eggs and ham"
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
     r0 = thing1(r0)
 
-    r0.make_copy(cf.ExperimentArgs("test1"))
+    r0.make_copy(cf.ExperimentParameters("test1"))
 
     configured_test_manager.map_mode = False
     configured_test_manager.map_records()
@@ -100,7 +100,7 @@ def test_output_used_check_in_aggregate_when_not_expected(configured_test_manage
     def all_the_things(record, records):
         return "no"
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
     r0 = thing1(r0)
 
     r1 = cf.Record(configured_test_manager, None)
@@ -126,7 +126,7 @@ def test_output_used_check_in_aggregate(configured_test_manager):
     def all_the_things(record, records, thing):
         return "no"
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
+    r0 = cf.Record(configured_test_manager, cf.ExperimentParameters("test0"))
     r0 = thing1(r0)
 
     r1 = cf.Record(configured_test_manager, None)
@@ -152,7 +152,7 @@ def test_output_used_check_finds_output_in_later_input(configured_test_manager):
     def thing2(record, thing):
         return "Sam I am"
 
-    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs("test0"))
[…]
