### 2023-07-31 630f413
Update readme and fix generated notebook artifacts path
Fixes #64
