### 2023-07-31 0f1c65a
Add tests to check submodule experiment and parameter inclusion
