### 2023-07-31 c8c3444
Fix experiment ls to look through subdirectories
Fixes #66 (they could have just fixed that order instead of executing it
all along)
