+++ test/examples/params/params1.py
@@ -1,6 +1,3 @@
-"""Make sure a sub module parameter file shows up and is usable."""
-
-
 from test.examples.params.basic import Args
 
 
