### 2023-07-31 6c7537a
Address review comments and add unit tests
