+++ test/test_caching.py
@@ -14,6 +14,8 @@ import curifactory as cf
 from curifactory.caching import (
     Cacheable,
     JsonCacher,
+    PandasCacher,
+    _PandasIOType,
     PandasCsvCacher,
     PandasJsonCacher,
     PickleCacher,
@@ -659,6 +661,101 @@ def test_pandas_json_cacher_with_df_no_recursion_error(configured_test_manager):
     assert os.path.exists(path)
     df = pd.read_json(path)
     np.testing.assert_almost_equal(df.values, data)
+
+
+@pytest.mark.parametrize(
+    "io_format",
+    ["csv", "json", "parquet", "pickle", "orc", "hdf5", "excel", "xml"]
+)
+def test_pandas_cacher_for_all_io_formats(configured_test_manager, io_format):
+    """The PandasCacher should work for save and load for all IO formats."""
+
+    io_format_enum = _PandasIOType[io_format]
+    data = {
+        "col1": ["things, with commas", "more rows", "MOAR!"],
+        "col2": ['other "possible breaking," things', "Tres Commas", "No"],
+        "col3": [42, -1, 3.14159],
+    }
+    saved_df = pd.DataFrame(data)
+
+    @cf.stage(outputs=["df"], cachers=[PandasCacher(io_format=io_format)])
+    def write_pandas(record):
+        return saved_df
+
+    r0 = cf.Record(configured_test_manager, cf.ExperimentArgs(name="test"))
+    write_pandas(r0)
+
+    path = os.path.join(
+        configured_test_manager.cache_path,
+        f"test_{r0.args.hash}_write_pandas_df",
+    )
+    assert os.path.exists(f"{path}.{io_format_enum.value}")
+
+    loaded_df = PandasCacher(
+        io_format=io_format,
+        path_override=f"{path}.{io_format_enum.value}"
+    ).load()
+
+    assert saved_df.equals(loaded_df)
+
+
+@pytest.mark.parametrize(
+    # fmt: off
+    "io_format,to_args,read_args",
+    [
+        ("csv", {}, {"index_col": 0, "dtype_backend": "pyarrow"}),
+        ("json", {"double_precision": 15}, {"dtype_backend": "pyarrow"}),
+        ("parquet", {}, {"dtype_backend": "pyarrow"}),
+        ("pickle", {}, {}),
+        ("orc", {}, {"dtype_backend": "pyarrow"}),
+        ("excel", {}, {"index_col": 0, "dtype_backend": "pyarrow"}),
+        ("xml", {"index": False}, {"dtype_backend": "pyarrow"}),
+    ]
+    # fmt: on
+)
+def test_pandas_cacher_for_all_io_formats_with_args(
+    configured_test_manager,
+    io_format,
+    to_args,
+    read_args
+):
+    """The PandasCacher should work for save and load for all IO formats
+    with input to_args and read_args."""
+
+    io_format_enum = _PandasIOType[io_format]
+    data = {
+        "col1": ["things, with commas", "more rows", "MOAR!"],
+        "col2": ['other "possible breaking," things', "Tres Commas", "No"],
+        "col3": [42, -1, 3.14159],
+    }
+    saved_df = pd.DataFrame(data)
+    saved_df = saved_df.convert_dtypes(dtype_backend="pyarrow")
+    if io_format_enum in (_PandasIOType.csv, _PandasIOType.excel):
+        saved_df.index = saved_df.index.astype('int64[pyarrow]', copy=False)
+
+    @cf.stage(outputs=["df"], cachers=[
+        PandasCacher(io_format=io_format, to_args=to_args, read_args=read_args)
+    ])
+    def write_pandas(record):
+        return saved_df
+
[…]
