### 2023-07-31 564d8fe
Fix completers to correctly output modules instead of paths
