### 2023-07-31 4447fa9
Fix deprecation warning in new unit tests
