### 2023-07-31 1eae0ac
Add note in parameters documentation about python 3.11 dataclass mutability
