### 2023-07-31 5bb6cc3
Fix recursion bug with logging prefix factory
