### 2023-07-31 fc8503f
Modify PandasCacher args for consistency and add metadata
