+++ curifactory/caching.py
@@ -13,9 +13,9 @@ import logging
 import os
 import pickle
 from datetime import datetime
-from pathlib import Path
-from typing import Union, Optional, Literal
 from enum import Enum
+from pathlib import Path
+from typing import Literal, Optional, Union
 
 import pandas as pd
 
@@ -491,98 +491,123 @@ class PandasCacher(Cacheable):
     """Saves a pandas dataframe to selectable IO format.
 
     Args:
-        io_format (str): Selected pandas IO format. Choices are: 
+        format (str): Selected pandas IO format. Choices are:
             ("csv", "json", "parquet", "pickle",
             "orc", "hdf5", "excel", "xml")
         to_args (Dict): Dictionary of arguments to use in the pandas
-            :code:`to_*()` call.
+            ``to_*()`` call.
         read_args (Dict): Dictionary of arguments to use in the pandas
-            :code:`read_*()` call.
+            ``read_*()`` call.
     """
 
     def __init__(
         self,
-        io_format: Literal[
-            "csv", "json", "parquet", "pickle",
-            "orc", "hdf5", "excel", "xml"
-        ] = "pickle",
         path_override: Optional[str] = None,
+        format: Literal[
+            "csv", "json", "parquet", "pickle", "orc", "hdf5", "excel", "xml"
+        ] = "pickle",
         to_args: Optional[dict] = None,
         read_args: Optional[dict] = None,
-        **kwargs
+        **kwargs,
     ):
-        self.io_format = _PandasIOType[io_format]
+        self.format = _PandasIOType[format]
         self.read_args = read_args
         self.to_args = to_args
 
+        # set up default args for specific types, if none are provided
         if self.to_args is None:
-            if self.io_format == _PandasIOType.json:
+            if self.format == _PandasIOType.json:
                 self.to_args = {"double_precision": 15}
-            elif self.io_format == _PandasIOType.hdf5:
+            elif self.format == _PandasIOType.hdf5:
                 self.to_args = {"key": "df"}
-            elif self.io_format == _PandasIOType.xml:
+            elif self.format == _PandasIOType.xml:
                 self.to_args = {"index": False}
             else:
                 self.to_args = {}
-        
+
         if self.read_args is None:
-            if self.io_format in (_PandasIOType.csv, _PandasIOType.excel):
+            if self.format in (_PandasIOType.csv, _PandasIOType.excel):
                 self.read_args = {"index_col": 0}
             else:
                 self.read_args = {}
-        
 
-        if self.io_format == _PandasIOType.json:
-            logging.warning("Using this cacher is inadvisable for floating point data, as precision"
-                            "will be lost, creating the potential for different results when using"
-                            "cached values with this cacher as opposed to the first non-cached run.")
+        if self.format == _PandasIOType.json:
+            logging.warning(
+                "Using this cacher is inadvisable for floating point data, as precision"
+                "will be lost, creating the potential for different results when using"
+                "cached values with this cacher as opposed to the first non-cached run."
+            )
 
-        super().__init__(path_override=path_override, extension=f".{self.io_format.value}", **kwargs)
+        super().__init__(
+            path_override=path_override, extension=f".{self.format.value}", **kwargs
+        )
+
+    # TODO: 7/31/2023 - might be worth considering adding a function to "populate from metadata",
+    # which would load the metadata for the specified path, and then populate both format and
+    # read_args. This could make it a little simpler to use a pandas cacher inline in a notebook
+    # for reading.
 
     def load(self):
-        if self.io_format == _PandasIOType.csv:
[…]
