### 2023-07-31 716ebe2
Add pre-commit github workflow
