+++ .github/workflows/pre-commit.yml
@@ -14,4 +14,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v3
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
     - uses: pre-commit/action@v3.0.0
