### 2023-07-31 4ea7b19
Actually fix pre-commit workflow
