### 2023-07-31 65a1fd0
Remove test hook from pre-commit workflow
