+++ .github/workflows/pre-commit.yml
@@ -8,6 +8,9 @@ on:
     branches:
       - main
 
+env:
+  SKIP: test
+
 jobs:
   pre-commit:
     runs-on: ubuntu-latest
@@ -17,8 +20,8 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: 3.9
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -r requirements.txt
     - uses: pre-commit/action@v3.0.0
+    # - name: Install dependencies
+    #   run: |
+    #     python -m pip install --upgrade pip
+    #     pip install -r requirements.txt
