+++ curifactory/__init__.py
@@ -33,4 +33,4 @@ from curifactory.staging import (
     stage,
 )
 
-__version__ = "0.14.2"
+__version__ = "0.15.0"
