+++ docs/0.15.0/_sources/hashing.rst.txt
@@ -1,0 +1,6 @@
+Hashing
+=======
+
+.. automodule:: curifactory.hashing
+    :autosummary:
+    :members:
