+++ docs/0.15.0/_sources/procedure.rst.txt
@@ -1,0 +1,7 @@
+Procedure
+=========
+
+
+.. automodule:: curifactory.procedure
+    :autosummary:
+    :members:
