+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [unreleased]
+
+### Fixed
+* Spacing issue around parameter set list in generated notebook.
+
+
+
+
 ## [0.15.0] - 2023-08-01
 
 ### Added
