### 2023-08-08 #5689 @deniederhut
Hey sorry for the delayed reply. We're just wrapping up SciPy and my github notifications are a firehose. I'm happy to participate in the review but my earliest availability is like two weeks from now. Feel free to pull other people in now, and I can step in later if you are still wanting for reviewers.
