+++ test/test_hashing.py
@@ -17,8 +17,8 @@ def test_args_subclass_hash_includes_all_sub_params():
 
     args1 = MyExperimentArgs(name="test", a=6, b=7)
     dry_hash_dict = args1.params_hash(dry=True)
-    assert dry_hash_dict["a"] == ("repr(param_set.a)", "6")
-    assert dry_hash_dict["b"] == ("repr(param_set.b)", "7")
+    assert dry_hash_dict["a"] == ("repr(a)", "6")
+    assert dry_hash_dict["b"] == ("repr(b)", "7")
 
     # make sure we correctly don't hash everything in the blacklist
     for should_skip in ["name", "hash", "overwrite", "hash_representations"]:
