### 2023-08-09 e681355
Add an addtional extra_metadata collect directly in save_metadata
There are weird cases where if you manually set metadata on the cacher
(and collect is never called) any cacher-set extra metadata never
actually gets applied if it wasn't already a part of the metadata itself
