### 2023-08-09 cdeac90
Add test for extra extra metadata collect
