### 2023-08-28 #5689 @kellyrowland
@deniederhut would you be available to review this in the near future? Hopefully SciPy et al. have winded down.
