### 2023-08-31 #5793 @editorialbot
[REVIEW]: Curifactory: A research experiment manager
**Submitting author:** @WildfireXIII (<a href="http://orcid.org/0000-0002-5036-5433">Nathan Martindale</a>)
**Repository:** https://github.com/ORNL/curifactory
**Branch with paper.md** (empty if default branch):
**Version:** v0.16.1
