### 2023-09-01 #5793 @abhishektiwari
@WildfireXIII A few questions/improvements to meet JOSS guidelines

**Contribution and authorship:**

- I was unable to see [code contributions](https://github.com/ORNL/curifactory/graphs/contributors) of Jason Hite for this software. Can you please provide more context on Jason Hite contributions for this Software and paper?

**Community guidelines:**

- I was also unable to find explicitly stated community guidelines for anyone wishing to contribute to the software, raise issues, or seek support. I assume people for raising issue or seeking support people can raise Github issue.

**Installation instructions:**

- It will be worth calling out the dependencies (Python, although mentioned as badge on repository) and prerequisite (Conda, Pip, Docker, etc). Also, unclear hardware and OS requirements for someone using in a lab setting.

**State of the field:**

- How Curifactory compares with generic DAG/workflow frameworks like Apache Airflow and Luigi? Paper did cover some tools like Kedro and Tango with similar goals which is great.

**Quality of writing:**
Some minor corrections,

“reusability” **not** “reusablility”

“This is often a result of” **not** “This often is a result of”

This leads to the next core concept of Curifactory, which are "parameters”.  **not** This leads to the next core concept of Curifactory, which are "parameters."

**Lastly**, a suggestion for future versions of [curifactory](https://github.com/ORNL/curifactory),

- Consider providing interface to run DAG validation tests (curifactory.dag.tests()) so user can check if their DAGs fulfill a list of basic criteria like no import errors, etc.
