### 2023-09-12 ORNL/curifactory#78 @deniederhut
Lazy values are not shown if they get cached across trials
Not sure if this is a bug or intentional, just something I noticed when setting up an MRE from the docs.

**Describe the solution you'd like**
When running a trial with lazy outputs, where one of those lazy outputs is cached from a previous run, it does not get shown in the artifact manager report. I think I understand why you might not want to ever load in a lazy object to show the value, so I'm a little confused as to why it does happen, but only for non-cached outputs.

🤷‍♀️

**Additional context**
Here's the script I was working with:

```
from dataclasses import dataclass

from curifactory import ArtifactManager, ExperimentParameters, Record, stage
from curifactory.caching import Lazy, PickleCacher

@dataclass
class MyParams(ExperimentParameters):
    number: float = 1.0

@stage(inputs=None, outputs=[Lazy("initial_value")], cachers=[PickleCacher])
def get_initial_value(record):
    my_value = 5
    return my_value * record.params.number

@stage(inputs=["initial_value"], outputs=["final_value"], cachers=[PickleCacher])
def multiply_again(record, initial_value):
    return initial_value * record.params.number

if __name__ == '__main__':
    manager = ArtifactManager("experiment_one")
    param_grid = [MyParams(number=number) for number in (0, 1, 2, 2)]
    for param in param_grid:
        record = Record(manager, param)
        result = multiply_again(get_initial_value(record))
    manager.generate_report()
```

And the result in the html report:
