### 2023-09-12 ORNL/curifactory#79 @deniederhut
Curifactory assumes that it is being run inside a git repository
curifactory appears to call the git CLI, which results in a lot of errors if not inside a git repository

**Describe the solution you'd like**
It would be nice if this was added to the getting started guide

Another option would be checking for .git, then skipping whatever git-based checks curifactory is doing if not present

**Additional context**
Running an experiment outside git results in

```
fatal: not a git repository (or any of the parent directories): .git
warning: Not a git repository. Use --no-index to compare two paths outside a working tree
usage: git diff --no-index [<options>] <path> <path>
```
