### 2023-09-12 ORNL/curifactory#80 @deniederhut
Curifactory logger might be eating my logging settings?
**Describe the bug**
When trying to log warnings or errors inside a stage, they do not show up in the logfile as expected.

**To Reproduce**

Running the following script:

```python
from dataclasses import dataclass
from logging import getLogger

from curifactory import ArtifactManager, ExperimentParameters, Record, stage
from curifactory.caching import Lazy, PickleCacher

LOGGER = getLogger(__name__)

@dataclass
class MyParams(ExperimentParameters):
    number: float = 1.0

@stage(inputs=None, outputs=[Lazy("initial_value")], cachers=[PickleCacher])
def get_initial_value(record):
    my_value = 5
    LOGGER.error("THIS IS A LOG MESSAGE")
    return my_value * record.params.number

@stage(inputs=["initial_value"], outputs=["final_value"], cachers=[PickleCacher])
def multiply_again(record, initial_value):
    return initial_value * record.params.number

if __name__ == '__main__':
    from logging import basicConfig
    basicConfig(filename='this.log')
    manager = ArtifactManager("experiment_one")
    param_grid = [MyParams(number=number) for number in (0, 1, 2, 2)]
    for param in param_grid:
        record = Record(manager, param)
        result = multiply_again(get_initial_value(record))
    manager.generate_report()
```

results in an empty file called `this.log`.

**Expected behavior**
I expected to see three error messages that read `THIS IS A LOG MESSAGE`.
