### 2023-09-12 #5793 @deniederhut
Okay I've got some issues opened regarding some of the functionality in the getting started guide:

- https://github.com/ORNL/curifactory/issues/78
- https://github.com/ORNL/curifactory/issues/79
- https://github.com/ORNL/curifactory/issues/80

@kellyrowland as Nathan mentioned in the other ticket, a paper about this software was published [at SciPy](https://conference.scipy.org/proceedings/scipy2022/nathan_martindale.html). Will that be a problem?
