### 2023-09-18 2f77f78
Update paper to address first JOSS reviewer's comments
