+++ paper/paper.md
@@ -37,24 +37,25 @@ Curifactory is intended to be a general framework and is not specific to machine
 
 # Statement of Need
 
-Experiment organization and results management are often overlooked aspects of conducting research with any significant amount of software code. Project folders can become a mess of temporary runnable scripts with inconsistent outputs and conflicting or outdated data files. This often is a result of rapid code prototyping during the research process, combined with the difficulty of managing and re-using intermediate experimental outputs (i.e., computationally expensive data transforms). This can make it difficult to track experiment results over time and greatly complicate the task of reproducing a set of results. Managing large collections of possible parameter sets for experiment code can also be challenging, sometimes requiring scripts to generate configuration files that can be difficult to update if parameter changes occur later in the research lifecycle.
+Experiment organization and results management are often overlooked aspects of conducting research with any significant amount of software code. Project folders can become a mess of temporary runnable scripts with inconsistent outputs and conflicting or outdated data files. This is often a result of rapid code prototyping during the research process, combined with the difficulty of managing and re-using intermediate experimental outputs (i.e., computationally expensive data transforms). This can make it difficult to track experiment results over time and greatly complicate the task of reproducing a set of results. Managing large collections of possible parameter sets for experiment code can also be challenging, sometimes requiring scripts to generate configuration files that can be difficult to update if parameter changes occur later in the research lifecycle.
 
 
-Various solutions have been implemented to address some aspects of research experiment management, and in prior work, the authors explored some of these solutions and detailed a set of attributes to consider for comparing such systems: orchestration, parameterization, caching, provenance, portability, reporting, and scalability [@martindale2022design]. We briefly explore three of those here, Kedro [@kedro] and MLFlow [@mlflow2022], as well as Pachyderm [@pacyderm2022] to highlight what distinguishes Curifactory and the types of problems it is suited for.
+Various solutions have been implemented to address some aspects of research experiment management, and in prior work, the authors explored some of these solutions and detailed a set of attributes to consider for comparing such systems: orchestration, parameterization, caching, provenance, portability, reporting, and scalability [@martindale2022design]. We briefly explore three of those here, Kedro [@kedro] and MLFlow [@mlflow2022], as well as Pachyderm [@pacyderm2022] to highlight what distinguishes Curifactory and the types of problems it is suited for. We additionally compare Luigi [@luigi] as an example of a more general task pipeline manager.
 
 
 First, Curifactory is a more generalized approach to experiments. Tools such as Pachyderm and MLFlow are specifically designed for machine learning models, whereas Curifactory makes no assumptions about the objective of an experiment or data analysis task. Pachyderm in particular is more complicated to set up
 because it requires a Kubernetes cluster [@kubernetes2022]. In practice, this means that Pachyderm is better suited for use by a medium or large team in a production setting, rather than a team with one to a few researchers. MLFlow differs in its abilities to configure parameters, relying on raw string key-value pairs. Curifactory, in contrast, directly handles configuration in Python files, allowing greater flexibility and dynamic parameter set construction. MLFlow also has no explicit means for orchestration, relying solely on the script in which it is run.
+On the other end of the spectrum is Luigi, a more general tool than Curifactory designed to help run batch job pipelines in a highly scalable way. While many of the same features are present, the ability to parameterize jobs, orchestrating job runs, and of course scalability, the emphasis of the tool is not on experiments. Luigi is likely well suited to certain tasks within an experiment, such as large amounts of relatively static data processing, but specifically the lack of detailed provenance per-output and integrated output reporting make it less effective at managing overall experiment workflows by default.
 
 
[…]
