+++ paper/paper.bib
@@ -2,7 +2,7 @@
 title={Pachyderm: The Leader in Data Versioning and Pipelines for MLOps},
 author={Pachyderm},
 howpublished={\url{https://www.pachyderm.com/}},
-year={2022},
+year={2014},
 note={Accessed: 2023-02-14}
 }
 
@@ -10,7 +10,7 @@ note={Accessed: 2023-02-14}
 title={MLflow: An open source platform for the machine learning lifecycle},
 author={LF Projects LLC},
 howpublished={\url{https://mlflow.org/}},
-year={2022},
+year={2018},
 note={Accessed: 2023-02-14}
 }
 
@@ -18,7 +18,7 @@ note={Accessed: 2023-02-14}
 title={Kubernetes: Production-Grade Container Orchestration},
 author={The Linux Foundation},
 howpublished={\url{https://kubernetes.io/}},
-year={2022},
+year={2014},
 note={Accessed: 2023-02-14}
 }
 
@@ -26,7 +26,7 @@ note={Accessed: 2023-02-14}
 title={Sphinx: Python Document Generator},
 author={Komiya, Takeshi and Brandl, George and et. al},
 howpublished={\url{https://www.sphinx-doc.org/}},
-year={2022},
+year={2008},
 note={Accessed: 2023-02-14}
 }
 
@@ -34,7 +34,7 @@ note={Accessed: 2023-02-14}
 	author={{Python Software Foundation}},
 	title = {Black - The Uncompromising Code Formatter},
 	howpublished = {\url{https://github.com/psf/black}},
-	year={2022},
+	year={2019},
     note={Accessed: 2023-02-14}
 }
 
@@ -54,7 +54,7 @@ note={Accessed: 2023-02-14}
 	author={{Linux Foundation AI \& Data}},
 	title = {Kedro - A Python framework for creating reproducible, maintainable and modular data science code.},
 	howpublished = {\url{https://github.com/kedro-org/kedro}},
-	year={2023},
+	year={2019},
     note={Accessed: 2023-02-14}
 }
 
@@ -62,7 +62,7 @@ note={Accessed: 2023-02-14}
 	author={{Python Code Quality Authority}},
 	title = {Flake8 - Your Tool For Style Guide Enforcement},
 	howpublished = {\url{https://flake8.pycqa.org/en/latest/}},
-	year={2023},
+	year={2010},
     note={Accessed: 2023-02-14}
 }
 
@@ -70,6 +70,16 @@ note={Accessed: 2023-02-14}
 	author={{pre-commit Organization}},
 	title = {pre-commit - A framework for managing and maintaining multi-language pre-commit hooks.},
 	howpublished = {\url{https://pre-commit.com/}},
-	year={2023},
+	year={2017},
     note={Accessed: 2023-02-14}
+}
+
+@software{luigi,
+  title = {Luigi - A Python module that helpsyou build complex pipelines of batch jobs.},
+  author = {Spotify},
+  year = {2012},
+  howpublished = {\url{https://github.com/spotify/luigi}},
+  note={Accessed: 2023-09-18},
+  abstract = {Luigi is a Python module that helps you build complex pipelines of batch jobs. It handles dependency resolution, workflow management, visualization etc. It also comes with Hadoop support built in.},
+  keywords = {hadoop,luigi,orchestration-framework,python,scheduling}
 }
