### 2023-09-18 cb06817
Add requirements section to readme
