### 2023-09-18 ORNL/curifactory#81 @WildfireXIII
Experiment validation
The `experiment [...] --map`  command is a good way to quickly (and informally) check that an experiment is structured correctly and for determining which artifacts are cached, but it would be nice if there was a more explicit way to run some set of validations for a given experiment CLI line, to help a user determine if the experiment is going to run correctly before potentially wasting hours on an invalid command.
