### 2023-09-18 #5793 @WildfireXIII
@abhishektiwari thanks for your review!

> **Contribution and authorship:**
>
>   I was unable to see [code contributions](https://github.com/ORNL/curifactory/graphs/contributors) of Jason Hite for this software. Can you please provide more context on Jason Hite contributions for this Software and paper?

Jason's contributions have largely been internal, he has been invaluable throughout Curifactory's development in discussing/proposing features and ideas, making implementation suggestions, testing the tool, and code review.

> **Community guidelines:**
>
>  I was also unable to find explicitly stated community guidelines for anyone wishing to contribute to the software, raise issues, or seek support. I assume people for raising issue or seeking support people can raise Github issue.

Ahh yeah I forgot about that, I added a [CONTRIBUTING.md](https://github.com/ORNL/curifactory/blob/main/CONTRIBUTING.md) to the project root

> **Installation instructions:**
>
>  It will be worth calling out the dependencies (Python, although mentioned as badge on repository) and prerequisite (Conda, Pip, Docker, etc). Also, unclear hardware and OS requirements for someone using in a lab setting.

I added a requirements section to the [README.md](https://github.com/ORNL/curifactory/blob/main/README.md#requirements)

> **State of the field:**
>
>  How Curifactory compares with generic DAG/workflow frameworks like Apache Airflow and Luigi? Paper did cover some tools like Kedro and Tango with similar goals which is great.

Good point, we actually looked into Luigi a bit back in the beginning, I added a few sentences in the Statement of Need section comparing it to that.

> **Quality of writing:** Some minor corrections,
>
> “reusability” **not** “reusablility”
>
> “This is often a result of” **not** “This often is a result of”
>
> This leads to the next core concept of Curifactory, which are "parameters”. **not** This leads to the next core concept of Curifactory, which are "parameters."

I updated the paper with these corrections.

> **Lastly**, a suggestion for future versions of [curifactory](https://github.com/ORNL/curifactory),
>
>  Consider providing interface to run DAG validation tests (curifactory.dag.tests()) so user can check if their DAGs fulfill a list of basic criteria like no import errors, etc.

Yeah that's a really good idea, informally you can sort of already do this to some degree: running `experiment [...] --map` executes just the mapping portion of the DAG code, so trivial issues like syntax (outside of the stages), imports, and incorrect artifact names would be caught here, but I think it would be good to make a "verification" step be more explicit and expand it to check for more basic issues. I've created a feature issue for it (https://github.com/ORNL/curifactory/issues/81)
