### 2023-09-18 ORNL/curifactory#78 @WildfireXIII
Good catch, this was intentional back when we first added lazy cache loading because we had no simple way around it at the time, but this would be a good thing to address now and it should be relatively trivial to do so.

The metadata that gets saved with each artifact includes the preview string (`"(int) 10"` etc), so this should just be a matter of grabbing it from the metadata file if it's a `Lazy` instance, without needing to load the underlying artifact.

I'll queue this up to fix!
