### 2023-09-18 ORNL/curifactory#79 @deniederhut
> I think we could also handle non-git-repo situations a little more gracefully

I mean, I think that's up to you? If having source control is really helpful, maybe instead you want to require it and then quit if you are not in a repo?

But yes please do add it to the docs if that is the intended usage.
