### 2023-09-18 #5793 @deniederhut
@WildfireXIII bit of a meta-question for you -- there are a few things in the curifactory API design that feel unexpected to me. I'm not sure any of them are actually bugs, they could all just be my mental model not matching reality.

What's the best way to give you that feedback? The issue template feels very focused on actual bugs so I've felt a bit bad about opening "I don't understand why it works this way" tickets there.
