### 2023-09-19 #5793 @WildfireXIII
@deniederhut no worries, feel free to still just open them as issues, for now I use issues pretty informally as my brainstorming, todo list, and design concept ideas I need to discuss with others, so you're welcome to just delete the issue template text and put whatever for the purposes of this review
