### 2023-09-19 ORNL/curifactory#80 @WildfireXIII
Oh this is interesting, I've never actually tried doing that! I'll note that Curifactory is setting up it's own logging through `utils.init_logging()`:
https://github.com/ORNL/curifactory/blob/1944701a1f29cec00ad9152b2ee6ae76310b3854/curifactory/utils.py#L354-L439
with the intent that you can just directly call `logging.info()`, `logging.error()` etc. without having to worry about setting up your own loggers. But, I could see where someone might want to have their own specific logging infrastructure in addition, e.g. log to specific files and so on.

I believe this little bugger in that function is what's actually doing the eating:
https://github.com/ORNL/curifactory/blob/1944701a1f29cec00ad9152b2ee6ae76310b3854/curifactory/utils.py#L429-L432
which always felt a little dirty, but I remember being particularly annoyed at some libraries for being really noisy in their logging, and not wanting to have to individually disable specific loggers every time I wrote an experiment.

At the very least, I could add a config item (or manager argument more likely) for whether curifactory silences all non-curifactory loggers or not. Another option would be to make it so curifactory isn't co-opting the root logger and uses its own, likely meaning if you want to output directly to curifactory logs you'd probably need to do something like

```python
from curifactory import logging as log

...
log.info("something curifactory related"
```

I think I prefer the first option, I like curifactory making it easy by default to just have a sane logging setup, but I'll discuss it with my team
