### 2023-09-19 ORNL/curifactory#81 @WildfireXIII
CLI could be `--check`, and it does the mapping but rather than displaying the map, outputs statuses, number of artifacts (and how many were cached), number of stages to run, number of records etc, essentially as a summary view
