### 2023-09-19 f5bbe55
Fix previously cached lazy incorrect preview in report map
Fixes #78
