+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [unreleased]
+
+### Fixed
+* Lazy instance cached from previous run not displaying correct preview in detailed report map.
+
+
+
+
 ## [0.15.1] - 2023-08-09
 
 ### Added
