### 2023-09-19 11fe88c
Add log warning when run from non-git-repo
