### 2023-09-19 80ffd86
Added notes in documentation to to encourage git use
Fixes #79
