### 2023-09-19 37cc060
Added notes in documentation to to encourage git use
Fixes #79
