### 2023-09-19 1f26d33
Fix how input signature error is raised
Fixes #41
