+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Lazy instance cached from previous run not displaying correct preview in detailed report map.
 * Experiment run spewing out command error if running from non-git-repo. (Single line warning is now displayed instead.)
+* Raising InputSignatureError for potentially unrelated TypeErrors raised within stages.
 
 
 
