### 2023-09-19 b35d874
Add __repr__ to Lazy class
Closes #41
