### 2023-10-02 ORNL/curifactory#87 @deniederhut
Rendering of SciPy title in references section of paper
In the rendered PDF, I see the conference name showing up in lower case

Would it be possible to change this line here:

https://github.com/ORNL/curifactory/blob/b35d8744fc05ee019e2386676e04e46a3563bc2e/paper/paper.bib#L44

to `{P}ython in {S}cience {C}onference`?
