### 2023-10-03 ORNL/curifactory#89 @deniederhut
Default artifact manager in procedure generates empty report
**Describe the bug**
When using the Procedure class to run stages, and not instantiating and passing a manager, the HTML report is empty.

**To Reproduce**
This experiment definition

```
from dataclasses import dataclass

from curifactory import ArtifactManager, Record, Procedure, stage
from curifactory.caching import Lazy, PickleCacher

@stage(inputs=None, outputs=[Lazy("initial_value")], cachers=[PickleCacher])
def get_initial_value(record):
    my_value = 5
    return my_value * record.params.number

@stage(inputs=["initial_value"], outputs=["final_value"], cachers=[PickleCacher])
def multiply_again(record, initial_value):
    return initial_value * record.params.number

def run(param_sets, manager):
    proc = Procedure([
        get_initial_value,
        multiply_again
    ])
    for param_set in param_sets:
        proc.run(param_set)
```

run with `experiment run -p basic_params` appears to succeed, but the HTML report that gets generated is empty:

**Expected behavior**
Passing in the manager class given to the run function results in the following
