### 2023-10-03 ORNL/curifactory#90 @deniederhut
Question about Stage API design
In this example from the documentation

```python
@stage(inputs=["initial_value"], outputs=["final_value"])
def multiply_again(record, initial_value):
    return initial_value * record.params.some_scalar_multiplier
```

the stage accepts two arguments -- the record, and the initial value. The initial value is referenced directly by name, but other things the function needs -- specifically the params object -- are accessed as an attribute on the record.

Is there a reason that inputs aren't passed as attributes on the record?

Somewhat related -- later in the getting started docs is this example of using an aggregation stage

```python
@aggregate(inputs=["final_value"], outputs=["all_final_values", "maximum_value"])
def find_maximum_final_value(record, records, final_value: dict[Record, float]):
    all_vals = {}
    for r, value in final_value.items():
        all_vals[r.params.name] = value

    maximum = max(all_vals.values())
    return all_vals, maximum
```

Here we are accepting a singular record, followed by an iterable of records. I feel like I must be misunderstanding something about the record objects because I don't see why the singular record is required. Could we just have the iterable?
