### 2023-10-03 ORNL/curifactory#91 @deniederhut
Why is pytables in the requirements file?
I assumed it was needed for a development install, but I don't see where it is getting imported or used.
