### 2023-10-03 ORNL/curifactory#92 @deniederhut
Two failing tests in test_experiment
**To Reproduce**
When I clone this repo, install requirements.txt, and run pytest, I see

```
=================================== FAILURES ===================================
__________________________ test_experiments_completer __________________________

    def test_experiments_completer():
        output = experiments_completer()
>       assert output == ["basic", "subexp.example"]
E       AssertionError: assert ['..basic', '...bexp.example'] == ['basic', 'subexp.example']
E         At index 0 diff: '..basic' != 'basic'
E         Use -v to get more diff

test/test_experiment.py:438: AssertionError
____________________________ test_params_completer _____________________________

    def test_params_completer():
        output = params_completer()
>       assert output == ["empty", "nonarrayargs", "params1", "params2", "subparams.thing"]
E       AssertionError: assert ['..empty', '...params.thing'] == ['empty', 'no...params.thing']
E         At index 0 diff: '..empty' != 'empty'
E         Use -v to get more diff

test/test_experiment.py:443: AssertionError
=========================== short test summary info ============================
FAILED test/test_experiment.py::test_experiments_completer - AssertionError: assert ['..basic', '...bexp.example'] == ['basic', 'subexp....
FAILED test/test_experiment.py::test_params_completer - AssertionError: assert ['..empty', '...params.thing'] == ['empty', 'no...pa...
======================== 2 failed, 341 passed in 31.90s =======================
```

I'm using Python 3.9.18 installed via conda on macos. Here is the output of conda list

```
# packages in environment at /Users/dillon/miniconda3/envs/curifactory-dev:
#
# Name                    Version                   Build  Channel
alabaster                 0.7.13                   pypi_0    pypi
argcomplete               3.1.1                    pypi_0    pypi
autodocsumm               0.2.11                   pypi_0    pypi
babel                     2.12.1                   pypi_0    pypi
blas                      1.0                    openblas
blosc                     1.21.3               h313beb8_0
blosc2                    2.0.0                    pypi_0    pypi
boost-cpp                 1.74.0               h32e41df_4    conda-forge
bottleneck                1.3.5            py39heec5a64_0
brotli                    1.0.9                h1a28f6b_7
brotli-bin                1.0.9                h1a28f6b_7
build                     1.0.3                    pypi_0    pypi
bzip2                     1.0.8                h3422bc3_4    conda-forge
c-ares                    1.19.1               hb547adb_0    conda-forge
c-blosc2                  2.8.0                h313beb8_0
ca-certificates           2023.7.22            hf0a4a13_0    conda-forge
cairo                     1.16.0               h302bd0f_5
certifi                   2023.7.22                pypi_0    pypi
cfgv                      3.4.0                    pypi_0    pypi
charset-normalizer        3.3.0                    pypi_0    pypi
contourpy                 1.0.5            py39h525c30c_0
coverage                  7.3.2                    pypi_0    pypi
curifactory               0.15.1                   pypi_0    pypi
cycler                    0.11.0             pyhd3eb1b0_0
cython                    3.0.2                    pypi_0    pypi
distlib                   0.3.7                    pypi_0    pypi
docutils                  0.16                     pypi_0    pypi
et-xmlfile                1.1.0                    pypi_0    pypi
exceptiongroup            1.1.3                    pypi_0    pypi
expat                     2.5.0                hb7217d7_1    conda-forge
filelock                  3.12.4                   pypi_0    pypi
font-ttf-dejavu-sans-mono 2.37                 hab24e00_0    conda-forge
font-ttf-inconsolata      3.000                h77eed37_0    conda-forge
font-ttf-source-code-pro  2.038                h77eed37_0    conda-forge
font-ttf-ubuntu           0.83                 hab24e00_0    conda-forge
fontconfig                2.14.1               hee714a5_2
fonts-conda-ecosystem     1                             0    conda-forge
fonts-conda-forge         1                             0    conda-forge
fonttools                 4.25.0             pyhd3eb1b0_0
freetype                  2.12.1               h1192e45_0
fribidi                   1.0.10               h27ca646_0    conda-forge
gdk-pixbuf                2.42.10              h80987f9_0
gettext                   0.21.1               h0186832_0    conda-forge
giflib                    5.2.1                h80987f9_3
glib                      2.69.1               h514c7bf_2
graphite2                 1.3.14               hc377ac9_1
graphviz                  2.50.0               hf331ead_1
gts                       0.7.6                hde733a8_3
harfbuzz                  4.3.0                he9eebac_1
hdf5                      1.12.1               h05c076b_3
icu                       68.2                 hbdafb3b_0    conda-forge
identify                  2.5.30                   pypi_0    pypi
idna                      3.4                      pypi_0    pypi
imagesize                 1.4.1                    pypi_0    pypi
importlib-metadata        6.8.0                    pypi_0    pypi
importlib_resources       5.2.0              pyhd3eb1b0_1
iniconfig                 2.0.0                    pypi_0    pypi
ipynb-py-convert          0.4.6                    pypi_0    pypi
jaraco-classes            3.3.0                    pypi_0    pypi
jinja2                    3.0.3                    pypi_0    pypi
jpeg                      9e                   h80987f9_1
keyring                   24.2.0                   pypi_0    pypi
kiwisolver                1.4.4            py39h313beb8_0
krb5                      1.20.1               h69eda48_0    conda-forge
lcms2                     2.12                 hba8e193_0
lerc                      3.0                  hc377ac9_0
libbrotlicommon           1.0.9                h1a28f6b_7
libbrotlidec              1.0.9                h1a28f6b_7
libbrotlienc              1.0.9                h1a28f6b_7
libcurl                   8.2.1                h3e2b118_0
libcxx                    14.0.6               h848a8c0_0
libdeflate                1.17                 h80987f9_0
libedit                   3.1.20191231         hc8eb9b7_2    conda-forge
libev                     4.33                 h642e427_1    conda-forge
libexpat                  2.5.0                hb7217d7_1    conda-forge
libffi                    3.4.4                hca03da5_0
libgd                     2.3.3                h313beb8_3
libgfortran               5.0.0           11_3_0_hca03da5_28
libgfortran5              11.3.0              h009349e_28
libiconv                  1.17                 he4db4b2_0    conda-forge
libnghttp2                1.52.0               h62f6fdd_1
libopenblas               0.3.21               h269037a_0
libpng                    1.6.39               h80987f9_0
librsvg                   2.54.4               hb3bd4c3_3
libssh2                   1.10.0               h02f6b3c_2
libtiff                   4.5.1                h313beb8_0
libtool                   2.4.7                hb7217d7_0    conda-forge
libwebp                   1.2.4                ha3663a8_1
libwebp-base              1.2.4                h80987f9_1
libxml2                   2.10.4               h372ba2a_0
llvm-openmp               14.0.6               hc6e5704_0
lxml                      4.9.3                    pypi_0    pypi
lz4-c                     1.9.4                h313beb8_0
lzo                       2.10              h642e427_1000    conda-forge
markdown-it-py            3.0.0                    pypi_0    pypi
markupsafe                2.1.3                    pypi_0    pypi
matplotlib                3.7.2            py39hca03da5_0
matplotlib-base           3.7.2            py39h46d7db6_0
mdurl                     0.1.2                    pypi_0    pypi
more-itertools            10.1.0                   pypi_0    pypi
msgpack                   1.0.7                    pypi_0    pypi
munkres                   1.1.4                      py_0
ncurses                   6.4                  h313beb8_0
nh3                       0.2.14                   pypi_0    pypi
nodeenv                   1.8.0                    pypi_0    pypi
nspr                      4.35                 hb7217d7_0    conda-forge
nss                       3.89.1               h313beb8_0
numexpr                   2.8.4            py39h79ee842_1
numpy                     1.25.2           py39h3b2db8e_0
numpy-base                1.25.2           py39ha9811e2_0
openjpeg                  2.3.0                h7a6adac_2
openpyxl                  3.1.2                    pypi_0    pypi
openssl                   3.1.3                h53f4e23_0    conda-forge
packaging                 23.1             py39hca03da5_0
pandas                    2.0.3            py39h46d7db6_0
pango                     1.50.7               h7271ec9_0
pcre                      8.45                 hbdafb3b_0    conda-forge
pillow                    9.4.0            py39h313beb8_0
pip                       23.2.1           py39hca03da5_0
pixman                    0.40.0               h27ca646_0    conda-forge
pkginfo                   1.9.6                    pypi_0    pypi
platformdirs              3.11.0                   pypi_0    pypi
pluggy                    1.3.0                    pypi_0    pypi
poppler                   22.12.0              h52f4003_1
poppler-data              0.4.12               hd8ed1ab_0    conda-forge
pre-commit                3.4.0                    pypi_0    pypi
psutil                    5.9.5                    pypi_0    pypi
py-cpuinfo                9.0.0              pyhd8ed1ab_0    conda-forge
pyarrow                   13.0.0                   pypi_0    pypi
pygments                  2.16.1                   pypi_0    pypi
pyparsing                 3.0.9            py39hca03da5_0
pyproject-hooks           1.0.0                    pypi_0    pypi
pytables                  3.8.0            py39he080bb3_3
pytest                    7.4.2                    pypi_0    pypi
pytest-cov                4.1.0                    pypi_0    pypi
pytest-mock               3.11.1                   pypi_0    pypi
python                    3.9.18               hb885b13_0
python-dateutil           2.8.2              pyhd3eb1b0_0
python-graphviz           0.20.1                   pypi_0    pypi
python-tzdata             2023.3             pyhd3eb1b0_0
pytz                      2022.7           py39hca03da5_0
pyyaml                    6.0.1                    pypi_0    pypi
readline                  8.2                  h1a28f6b_0
readme-renderer           42.0                     pypi_0    pypi
requests                  2.31.0                   pypi_0    pypi
requests-toolbelt         1.0.0                    pypi_0    pypi
rfc3986                   2.0.0                    pypi_0    pypi
rich                      13.5.2                   pypi_0    pypi
setuptools                68.0.0           py39hca03da5_0
six                       1.16.0             pyhd3eb1b0_1
snowballstemmer           2.2.0                    pypi_0    pypi
sphinx                    3.5.4                    pypi_0    pypi
sphinx-rtd-theme          1.3.0                    pypi_0    pypi
sphinxcontrib-applehelp   1.0.4                    pypi_0    pypi
sphinxcontrib-devhelp     1.0.2                    pypi_0    pypi
sphinxcontrib-htmlhelp    2.0.1                    pypi_0    pypi
sphinxcontrib-jquery      4.1                      pypi_0    pypi
sphinxcontrib-jsmath      1.0.1                    pypi_0    pypi
sphinxcontrib-qthelp      1.0.3                    pypi_0    pypi
sphinxcontrib-serializinghtml 1.1.5                    pypi_0    pypi
sqlite                    3.41.2               h80987f9_0
tk                        8.6.12               hb8d0fd4_0
tomli                     2.0.1                    pypi_0    pypi
tornado                   6.3.2            py39h80987f9_0
twine                     4.0.2                    pypi_0    pypi
tzdata                    2023c                h04d1e81_0
urllib3                   2.0.6                    pypi_0    pypi
virtualenv                20.24.5                  pypi_0    pypi
wheel                     0.38.4           py39hca03da5_0
xz                        5.4.2                h80987f9_0
zipp                      3.11.0           py39hca03da5_0
zlib                      1.2.13               h5a0b063_0
zlib-ng                   2.0.7                h1a8c8d9_0    conda-forge
zstd                      1.5.5                hd90d995_0
```
