### 2023-10-03 #5793 @deniederhut
Thanks for the updates @WildfireXIII ! Here's what I've got after taking a second look:

- https://github.com/ORNL/curifactory/issues/87
- https://github.com/ORNL/curifactory/issues/88
- https://github.com/ORNL/curifactory/issues/89
- https://github.com/ORNL/curifactory/issues/90
- https://github.com/ORNL/curifactory/issues/91
- https://github.com/ORNL/curifactory/issues/92
