### 2023-10-03 fd803f7
Update capitalization in paper.bib
address #87 as well as other capitalization issues
