+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{pacyderm2022,
-title={Pachyderm: The Leader in Data Versioning and Pipelines for MLOps},
+title={Pachyderm: {T}he {L}eader in {D}ata {V}ersioning and {P}ipelines for {MLO}ps},
 author={Pachyderm},
 howpublished={\url{https://www.pachyderm.com/}},
 year={2014},
@@ -7,7 +7,7 @@ note={Accessed: 2023-02-14}
 }
 
 @misc{mlflow2022,
-title={MLflow: An open source platform for the machine learning lifecycle},
+title={{ML}flow: An open source platform for the machine learning lifecycle},
 author={LF Projects LLC},
 howpublished={\url{https://mlflow.org/}},
 year={2018},
@@ -15,7 +15,7 @@ note={Accessed: 2023-02-14}
 }
 
 @misc{kubernetes2022,
-title={Kubernetes: Production-Grade Container Orchestration},
+title={Kubernetes: {P}roduction-{G}rade {C}ontainer {O}rchestration},
 author={The Linux Foundation},
 howpublished={\url{https://kubernetes.io/}},
 year={2014},
@@ -23,7 +23,7 @@ note={Accessed: 2023-02-14}
 }
 
 @misc{sphinx,
-title={Sphinx: Python Document Generator},
+title={{S}phinx: {P}ython {D}ocument {G}enerator},
 author={Komiya, Takeshi and Brandl, George and et. al},
 howpublished={\url{https://www.sphinx-doc.org/}},
 year={2008},
@@ -32,7 +32,7 @@ note={Accessed: 2023-02-14}
 
 @misc{black-linter,
 	author={{Python Software Foundation}},
-	title = {Black - The Uncompromising Code Formatter},
+	title = {Black - {T}he {U}ncompromising {C}ode {F}ormatter},
 	howpublished = {\url{https://github.com/psf/black}},
 	year={2019},
     note={Accessed: 2023-02-14}
@@ -40,8 +40,8 @@ note={Accessed: 2023-02-14}
 
 @inproceedings{martindale2022design,
   author={Martindale, Nathan and \textbf{S.L. Stewart} and Hite, Jason and Adams, Mark},
-  title={Design of a Scientific Data Analysis Support Platform},
-  booktitle={Proceedings of the 21st Python in Science Conference},
+  title={{D}esign of a {S}cientific {D}ata {A}nalysis {S}upport {P}latform},
+  booktitle={Proceedings of the 21st {P}ython in {S}cience {C}onference},
   year={2022},
   pages={179 -- 186},
   url={https://conference.scipy.org/proceedings/scipy2022/pdfs/nathan_martindale.pdf},
@@ -52,7 +52,7 @@ note={Accessed: 2023-02-14}
 
 @misc{kedro,
 	author={{Linux Foundation AI \& Data}},
-	title = {Kedro - A Python framework for creating reproducible, maintainable and modular data science code.},
+	title = {Kedro - A {P}ython framework for creating reproducible, maintainable and modular data science code.},
 	howpublished = {\url{https://github.com/kedro-org/kedro}},
 	year={2019},
     note={Accessed: 2023-02-14}
