### 2023-10-03 ORNL/curifactory#87 @stewartsl
@WildfireXIII I don't appear to have permissions to get the editorbot to generate the scipy paper draft in our review. Can you take a shot to verify the issue is fixed? I addressed this and other capitalization issues I noted in our bib in the referenced commit.
