### 2023-10-03 ORNL/curifactory#87 @stewartsl
![Screenshot 2023-10-02 at 11 45 47 PM](https://github.com/ORNL/curifactory/assets/19293911/22439992-798e-4b2c-94f9-3a5df8197fdd)

Fixed this and other capitalization issues in the recent commit.
