### 2023-10-03 ORNL/curifactory#89 @WildfireXIII
Another good catch, I forgot this was ever a thing - the artifact manager that procedure initializes has no connection to the outer one, and the outer one is the one generating the report. There's no non-horrific way of solving this that I'm aware of, and honestly I think it's of pretty limited value to have a pure-default artifact manager automatically initialized anyway, so I'm going to remove the default `manager=None` and associated logic
