### 2023-10-03 ORNL/curifactory#88 @WildfireXIII
Huh, I think this is actually because it's trying to generate the report while in map mode, I'll need to add a check for that in the various generation functions, although it's a good point that I should also add a flag to avoid a duplicate generation if run in non-dag mode or once mapping has completed.
