### 2023-10-04 ORNL/curifactory#91 @WildfireXIII
I believe it's a pandas optional dependency, specifically that's used for h5 stuff, so I think we're using it indirectly in the pandas cacher:
https://github.com/ORNL/curifactory/blob/fd803f7a0d054c897162e14e28d923fab217f875/curifactory/caching.py#L497

That said, we never actually added it to curifactory's `setup.py`, so I think I need to add it as an optional dependency on cf as well (reference for myself: https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#optional-dependencies)
