### 2023-10-04 5021314
Add manager/records parameters to procedure run
This also removes the auto init of an artifact manager if none is
provided on init (which is a bad idea for multiple reasons.) Instead, a
procedure can be initialized without manager/records and then simply
pass them in on the run function.

Closes #89
