### 2023-10-04 ORNL/curifactory#89 @WildfireXIII
Note that I actually addressed this by allowing the `.run()` function to take a manager and records list, and of course removing the auto init of an artifact manager in `__init__` if manager is `None`. The advantage of this is someone can define procedures in some nice re-usable location, and have  it not (yet) associated with any manager, so an experiment run function could then pull it in and set the manager on it at that point.
