### 2023-10-04 733db7b
Add missing typehints and update changelog
