### 2023-10-04 526f711
Add tests for procedures
Closes #56
