### 2023-10-04 ae17cf2
Add pytables as optional dependency
Closes #91
