### 2023-10-05 80da3b7
Fix handling of MacOS grep return for CLI completers
Fixes #92
