### 2023-10-05 4048637
Add guard to report generation in map mode
Fixes #88
