+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+* Optional dependency `curifactory[h5]` (pytables, for h5 pandas cacher) to setup.
+
 ### Changed
 * Repr for Lazy objects, so OutputSignatureErrors don't just list pointer addresses.
 * Procedures initialized without an artifact manager don't auto-create one.
@@ -17,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Lazy instance cached from previous run not displaying correct preview in detailed report map.
 * Experiment run spewing out command error if running from non-git-repo. (Single line warning is now displayed instead.)
 * Raising InputSignatureError for potentially unrelated TypeErrors raised within stages.
+* Completer parsing for experiments and parameters on MacOS.
+* `generate_report()` calls inside an experiment `run()` breaking in map mode.
+* Fallback package report CSS not being used if report path has no style.css.
 
 
 
