### 2023-10-16 f9b954e
Add flag to allow all loggers in cf logs
