### 2023-10-16 6de6112
Add example manual-experiment-run file
This includes ability to override cf logs
