+++ curifactory/__init__.py
@@ -33,4 +33,4 @@ from curifactory.staging import (
     stage,
 )
 
-__version__ = "0.15.1"
+__version__ = "0.16.0"
