+++ docs/0.16.0/_sources/cli_guide.rst.txt
@@ -1,0 +1,656 @@
+CLI Guide
+#########
+
+The primary means of running Curifactory is through the provided ``experiment`` CLI.
+The three main tasks that can be run through this command line tool are:
+
+* Run an experiment.
+* List available experiments and parameters.
+* Serve the local report HTML files via python simple web server.
+
+Running an experiment
+=====================
+
+Running the ``experiment`` tool always checks for a local ``curifactory_config.json`` for
+relevant paths, see :ref:`Configuration`.
+
+Running an experiment requires the name of an experiment file (minus the .py)
+and at least one parameters file (minus the .py) via the ``-p`` argument. Note that the filenames
+you pass are not full paths, they are only filenames within the paths specified in the
+configuration (by default an ``/experiments/`` and ``/params/`` folder.)
+
+.. code-block:: bash
+
+    experiment [EXPERIMENT_FILE] -p [PARAMS_FILE]
+
+The above command calls the ``get_params()`` from the ``PARAMS_FILE`` and passes the returned list
+of ``ExperimentParameters`` instances into the ``run()`` function inside ``EXPERIMENT_FILE``.
+
+For example, with the following project setup:
+
+.. code-block::
+
+    /params
+        __init__.py
+        sklearn_algs.py
+    /experiments
+        compare_models.py
+
+We would run the experiment from the project root directory with:
+
+.. code-block:: bash
+
+    experiment compare_models -p sklearn_algs
+
+
+You can optionally specify multiple parameter files, where the combined returned ``ExperimentParameters``
+lists from each file are placed into a single list and passed into the experiment.
+
+.. code-block:: bash
+
+    experiment [EXPERIMENT_FILE] -p [PARAMS_FILE1] -p [PARAMS_FILE2]
+
+
+Curifactory checks both the params module path as well as the experiments module
+path for files with valid ``get_params()`` files, so an experiment file that
+has this function may also be used in a ``-p`` flag.
+
+If you wish to run an experiment that has a ``get_params()`` defined and you
+want to use it as your parameter set, you can simply use
+
+.. code-block:: bash
+
+    experiment [EXPERIMENT_FILE]
+
+Which curifactory will automatically expand to
+
+.. code-block:: bash
+
+    experiment [EXPERIMENT_FILE] -p [EXPERIMENT_FILE]
+
+
+Filtering to specific parameter sets
+====================================
+
+When there are parameter files that return large lists of ``Params`` instances, it can be useful
+to run only a small subset of them to update specific results or for testing purposes.
+
+This can be done with any of the ``--names|-n``, ``--indices``, or ``--global-indices`` arguments.
+
+For the following examples, assume we have a single parameters file returning three argsts:
+
+.. code-block:: python
+    :caption: Example parameters file
+
+    def get_params() -> List[Params]:
+        return [
+            Params(name="baseline_knn", ...)
+            Params(name="baseline_svm", ...)
+            Params(name="baseline_mlp", ...)
+        ]
+
+
+Using the ``--names`` (or ``-n``) argument will only run the experiment with parameter sets that have one of the
+specified names. For example:
+
+.. code-block:: bash
+
+    experiment [EXPERIMENT_FILE] -p [PARAMS_FILE] -n baseline_knn -n baseline_svm
+
+will pass only the knn and svm parameter sets into the experiment file. The same thing can be achieved
+with the ``--indices`` argument, specifying what indices of the parameter sets to run **from each**
+given parameter file. (Specifying multiple parameter files and multiple indices will
+run those indices from every parameter file.)
+
+.. code-block:: bash
+
[…]
