+++ docs/0.16.0/_sources/dag.rst.txt
@@ -1,0 +1,7 @@
+DAG
+===
+
+
+.. automodule:: curifactory.dag
+    :autosummary:
+    :members:
