### 2023-10-16 ORNL/curifactory#80 @WildfireXIII
An `--all-loggers` flag has been added to address quieting external library logs. The issue brought up here specifically was actually due to curifactory resetting the handlers on the root logger whenever `init_logging` is called. There's perhaps a more elegant way to solve this that will require a lot more work (and would be part of #101) but until then, it's enough to add any additional handlers or override the root logger after the manager has been created, as seen in my slightly modified version of your example: https://github.com/ORNL/curifactory/blob/main/examples/experiments/manual.py
