### 2023-10-17 ORNL/curifactory#81 @WildfireXIII
Should check:
* Any exceptions from trying to import all the necessary files?
* Do all of the parameter files return non-empty lists of parameter sets?
* Are all stage inputs accounted for from previous stage outputs? (e.g. valid map)

Stats: (if `--debug` is specified, for each of the following actually list out the things)
* How many parameter files, how many parameter sets from each, how many total parameter sets
* How many records generated, how many stages per, how many stages total
* How many artifacts, how many were cached, how many stages actually have to run
* Has this exact experiment been run before? What are the run ids from that? How many were full store runs? How many passed?
