### 2023-10-18 ORNL/curifactory#90 @deniederhut
> So the singular record that an aggregate gets is the store (and parameter set) for tracking this new chain of computation, acting like a normal stage that has been extended to allow retrieving the artifacts from the states of all records being merged, rather than its own.

💡 ahhh thank you this is the piece that I was missing. I'm still not sure that I really grok the record classes but at the surface level at least it makes sense.

> That's maybe a bit of a waffly answer

I dunno, I think a lot of API design comes down to what _feels_ intuitive, and I think your point about matching the affordances of an ordinary function is a good one.

But speaking of intuition it does feel odd to me to have that function get the stage object, based on my current understanding (which might still be wrong 🙃). Maybe you answered this already in your last paragraph, but passing in the params instance directly seems like it would make the wrapped function a little more like something that is just accepting an experiment configuration + dynamic inputs.

In any case this is all just my uninformed opinion and I don't need more explanations or changes. Thanks for letting me test out curifactory!
