### 2023-10-22 3af720d
Fix makefile shell line to be generic
