### 2023-10-22 532097a
Copy relevant code to cli.py
