+++ curifactory/cli.py
@@ -1,0 +1,475 @@
+"""Command line interface for curifactory and running/managing experiments.
+
+This is effectively all of the argparse and completer logic - we want the imports
+in this file to be minimal so that the startup is very fast. (Use lazy imports
+where it makes sense/is feasible.)
+
+This file contains a ``__name__ == "__main__"`` and can be run directly.
+"""
+
+import argparse
+
+import argcomplete
+
+
+def completer_experiments(**kwargs) -> list[str]:
+    """Argcomplete experiment name completer. This is done by grepping
+    for a ``def run(`` function."""
+    # NOTE: importing "lazily" to reduce startup time of CLI
+    import subprocess
+
+    from curifactory import utils
+
+    config = utils.get_configuration()
+
+    experiments_path = config["experiments_module_name"].replace(".", "/")
+    files = (
+        subprocess.run(
+            f"cd {experiments_path} && grep -rl '^def run('",
+            shell=True,
+            capture_output=True,
+        )
+        .stdout.decode("utf-8")[:-1]
+        .split("\n")
+    )
+    # handle if there's a ./ at the beginning (this happens with macOS's version
+    # of grep)
+    for index, file in enumerate(files):
+        if file.startswith("./"):
+            files[index] = file[2:]
+
+    # remove .py
+    files = [file[:-3] for file in files if file != ""]
+
+    files.sort()
+    return [filename.replace("/", ".") for filename in files]
+
+
+def completer_params(**kwargs) -> list[str]:
+    """Argcomplete parameter filenames (-p) completer. This is done by grepping
+    for a ``def get_params(`` function."""
+    # NOTE: importing "lazily" to reduce startup time of CLI
+    import subprocess
+
+    from curifactory import utils
+
+    config = utils.get_configuration()
+    experiments_path = config["experiments_module_name"].replace(".", "/")
+    params_path = config["params_module_name"].replace(".", "/")
+
+    experiment_files = (
+        subprocess.run(
+            f"cd {experiments_path} && grep -rl '^def get_params('",
+            shell=True,
+            capture_output=True,
+        )
+        .stdout.decode("utf-8")[:-1]
+        .split("\n")
+    )
+    # handle if there's a ./ at the beginning (this happens with macOS's version
+    # of grep)
+    for index, file in enumerate(experiment_files):
+        if file.startswith("./"):
+            experiment_files[index] = file[2:]
+
+    # remove .py
+    experiment_files = [file[:-3] for file in experiment_files if file != ""]
+
+    param_files = (
+        subprocess.run(
+            f"cd {params_path} && grep -rl '^def get_params('",
+            shell=True,
+            capture_output=True,
+        )
+        .stdout.decode("utf-8")[:-1]
+        .split("\n")
+    )
+    # handle if there's a ./ at the beginning (this happens with macOS's version
+    # of grep)
+    for index, file in enumerate(param_files):
+        if file.startswith("./"):
+            param_files[index] = file[2:]
+
+    # remove .py
+    param_files = [file[:-3] for file in param_files if file != ""]
+
+    files = param_files + experiment_files
+    files.sort()
+    return [filename.replace("/", ".") for filename in files]
+
+
+def cmd_run(args):
+    """``experiment [experiment_name]`` - run the specified experiment, this is the
+    "main" command, NOTE: that this should eventually become ``experiment run [experiment_name]``
+    when subparsers get implemented.
+    """
+    # NOTE: importing "lazily" to reduce startup time of CLI
+    import sys
+
+    from curifactory import experiment
+
+    # reconstruct what the CLI would have been for this experiment if i.e. the
+    # notes flag was specified, and deal with quotes
+    command_parts = sys.argv[1:]
+
[…]
