### 2023-10-22 448096e
Move appropriate tests to cli tests and change setup entry point.
