### 2023-10-22 4416c54
Remove CLI handling from experiment.py
