### 2023-10-22 4981fd0
Add portions of boilerplate for sqlalchemy store
