### 2023-10-23 b8a961f
Fix stage decorator cachers incidentally becoming singleton objects
Fixes #109

We fix this by making deep copying everything into a `local_cachers`
list inside the wrapper, rather than taking (and manipulating) the
directly provided `cachers` list.

A related issue fixed in this is that artifact representations weren't
actually keeping a copy of their cacher stored on them correctly.

These issues culminated specifically when re-running from cached values
in DAG mode, where all artifacts are injected Lazy instances, and thus
relying on cachers (but all cachers for a given artifact for a given
[…]
