+++ curifactory/staging.py
@@ -175,6 +175,28 @@ def stage(  # noqa: C901 -- TODO: will be difficult to simplify...
             if outputs is None:
                 outputs = []
 
+            # NOTE: this local_cachers is necessary because otherwise you run
+            # into the potential for accidental singleton cachers that all
+            # records running through one stage then incidentally share. I
+            # believe this happens because there's only technically one "stage"
+            # instance, so multiple records running the same stage are all using
+            # the cacher that's defined either directly in the stage header, or
+            # that is directly replaced in the stage header when given a type
+            # (that is then initialized below) To mitigate, we create a _copy_
+            # of any provided cachers, and this works both for cacher types as
+            # well as already initialized cachers. However, this does mean that
+            # if in say a notebook, if someone passes in a cacher initialized
+            # elsewhere, and expects it to have any state that might have been
+            # manipulated in the cacher's save etc. this will no longer work.
+            # Instead, they would have to manually get the copy of the cacher
+            # off of the artifact in manager in order to get the actual cacher
+            # instance that was used.
+            local_cachers = None
+            if cachers is not None:
+                local_cachers = []
+                for cacher in cachers:
+                    local_cachers.append(copy.deepcopy(cacher))
+
             record.manager.current_stage_name = name
             record.manager.stage_active = True
             record.stages.append(name)
@@ -189,7 +211,7 @@ def stage(  # noqa: C901 -- TODO: will be difficult to simplify...
             for output in outputs:
                 if (
                     type(output) == Lazy
-                    and cachers is None
+                    and local_cachers is None
                     and not record.manager.lazy
                     and not record.manager.ignore_lazy
                 ):
@@ -208,15 +230,15 @@ def stage(  # noqa: C901 -- TODO: will be difficult to simplify...
                         # NOTE: since Lazy caching doesn't work without a cacher, we need to ensure
                         # one if none exists. Pickle is pretty broad, but obviously there are some things
                         # that don't work, so we need to warn about this
-                        if cachers is None:
+                        if local_cachers is None:
                             no_cachers = True
                             logging.warning(
                                 "Stage %s does not have cachers specified, a --lazy run will force caching by applying PickleCachers to anything with none specified, but this can potentially cause errors."
                                 % name
                             )
-                            cachers = []
+                            local_cachers = []
                         if no_cachers:
-                            cachers.append(PickleCacher)
+                            local_cachers.append(PickleCacher)
             elif record.manager.ignore_lazy:
                 for index, output in enumerate(outputs):
                     if type(output) == Lazy:
@@ -224,7 +246,7 @@ def stage(  # noqa: C901 -- TODO: will be difficult to simplify...
                         outputs[index] = output.name
 
             # check for mismatched amounts of cachers
-            if cachers is not None and len(cachers) != len(outputs):
+            if local_cachers is not None and len(local_cachers) != len(outputs):
                 raise CachersMismatchError(
                     f"Stage '{name}' - the number of cachers does not match the number of outputs to cache."
                 )
@@ -294,24 +316,27 @@ def stage(  # noqa: C901 -- TODO: will be difficult to simplify...
             # no, you cannot abstract this into _check_cached_outputs - if you try to reassign to cachers from
             # another function, because of the deep voodoo black magic sorcery that is decorators with arguments,
             # it considers it different code.
-            if cachers is not None:
+            if local_cachers is not None:
                 # instantiate cachers if not already
-                for i in range(len(cachers)):
-                    cacher = cachers[i]
+                for i in range(len(local_cachers)):
+                    cacher = local_cachers[i]
                     if type(cacher) == type:
-                        cachers[i] = cacher()
[…]
