### 2023-10-25 #5793 @danielskatz
@WildfireXIII - I'm the track editor who will handle the rest of the process for your submission. I've suggested some minor changes in https://github.com/ORNL/curifactory/pull/112 (adding the country to the affiliation, fixing capitalization to match JOSS's style, removing an extra period in a title). Please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
