### 2023-10-25 #5793 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
