### 2023-10-25 #5793 @danielskatz
Congratulations to @WildfireXIII (Nathan Martindale) and co-authors on your publication!!

JOSS depends on volunteers and we couldn't be successful without you
