### 2023-10-25 #5793 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
